[REFACTOR/#355] 프로필, 세팅 뷰모델 분리
